+++ lib/dotcom_web/controllers/health_controller.ex
@@ -2,13 +2,10 @@ defmodule DotcomWeb.HealthController do
   @moduledoc """
   Simple controller to return 200 OK when the website is healthy.
   """
-
   use DotcomWeb, :controller
 
   def index(conn, _params) do
-    status = gettext("All clear, captain!")
-
     conn
-    |> send_resp(200, status)
+    |> send_resp(200, "")
   end
 end
