+++ priv/gettext/dotcom.pot
@@ -1,0 +1,17 @@
+## This file is a PO Template file.
+##
+## "msgid"s here are often extracted from source code.
+## Add new messages manually only if they're dynamic
+## messages that can't be statically extracted.
+##
+## Run "mix gettext.extract" to bring this file up to
+## date. Leave "msgstr"s empty as changing them here has no
+## effect: edit them in PO (.po) files instead.
+#
+msgid ""
+msgstr ""
+
+#: lib/dotcom_web/controllers/test_controller.ex:11
+#, elixir-autogen, elixir-format
+msgid "All clear, captain!"
+msgstr ""
