+++ priv/gettext/es/LC_MESSAGES/dotcom.po
@@ -1,0 +1,17 @@
+## "msgid"s in this file come from POT (.pot) files.
+###
+### Do not add, change, or remove "msgid"s manually here as
+### they're tied to the ones in the corresponding POT file
+### (with the same domain).
+###
+### Use "mix gettext.extract --merge" or "mix gettext.merge"
+### to merge POT files into PO files.
+msgid ""
+msgstr ""
+"Language: es\n"
+"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+
+#: lib/dotcom_web/controllers/test_controller.ex:11
+#, elixir-autogen, elixir-format
+msgid "All clear, captain!"
+msgstr ""
